# docs/api/swagger.yaml
basePath: /api/v1
definitions:
  big.Int:
    type: object
...
  fmis-api_internal_models.TransactionType:
    enum:
    - INCOMING
    - OUTGOING
    - USED_IN_PRODUCTION
    - WASTE
    - ADJUSTMENT
    type: string
    x-enum-varnames:
    - TransactionTypeIncoming
    - TransactionTypeOutgoing
    - TransactionTypeUsedInProduction
    - TransactionTypeWaste
    - TransactionTypeAdjustment
  fmis-api_internal_schemas.AdjustStockRequest:
    properties:
      batch_id:
        type: string
      quantity_change:
        type: string
      reference_note:
        maxLength: 500
        type: string
      transaction_type:
        enum:
        - WASTE
        - ADJUSTMENT
        type: string
    required:
    - batch_id
    - quantity_change
    - transaction_type
    type: object
  fmis-api_internal_schemas.ConsumeStockRequest:
    properties:
      product_id:
        type: string
      quantity:
        type: string
      reference_note:
        maxLength: 500
        type: string
    required:
    - product_id
    - quantity
    type: object
...
  fmis-api_internal_schemas.StockTransaction:
    properties:
      batch_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      performed_by:
        type: string
      production_order_id:
        type: string
      quantity_change:
        $ref: '#/definitions/pgtype.Numeric'
      reference_note:
        $ref: '#/definitions/pgtype.Text'
      transaction_type:
        $ref: '#/definitions/fmis-api_internal_models.TransactionType'
    type: object
...
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    format: int32
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
  pgtype.Numeric:
    properties:
      exp:
        format: int32
        type: integer
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      int:
        $ref: '#/definitions/big.Int'
      naN:
        type: boolean
      valid:
        type: boolean
    type: object
  pgtype.Text:
    properties:
      string:
        type: string
      valid:
        type: boolean
    type: object
...
info:
  contact: {}
  description: Food Manufacturing Inventory System REST API.
  title: FMIS API
  version: 1.0.0
paths:
...
  /inventory/adjust:
    post:
      consumes:
      - application/json
      description: Records a WASTE or ADJUSTMENT stock movement on a single batch.
        ADMIN or OPERATOR role required.
      parameters:
      - description: Adjustment payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/fmis-api_internal_schemas.AdjustStockRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/fmis-api_internal_schemas.StockTransaction'
        "400":
          description: Invalid body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Missing or invalid token
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Insufficient role
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Batch not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Batch state prevents adjustment
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Quantity would exceed available stock
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Adjust or waste stock
      tags:
      - inventory
  /inventory/consume:
    post:
      consumes:
      - application/json
      description: Deducts stock across batches of a product, oldest expiration first
        (FEFO). ADMIN or OPERATOR role required.
      parameters:
      - description: Consumption payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/fmis-api_internal_schemas.ConsumeStockRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/fmis-api_internal_schemas.StockTransaction'
            type: array
        "400":
          description: Invalid body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Missing or invalid token
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Insufficient role
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Insufficient stock
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Consume stock (FEFO)
      tags:
      - inventory
  /inventory/transactions/:
    get:
      description: Returns the immutable audit trail of stock movements, optionally
        filtered.
      parameters:
      - description: Filter by batch UUID
        in: query
        name: batch_id
        type: string
      - description: Filter by type (INCOMING, OUTGOING, USED_IN_PRODUCTION, WASTE,
          ADJUSTMENT)
        in: query
        name: transaction_type
        type: string
      - description: Only transactions created at or after this RFC3339 timestamp
        in: query
        name: date_from
        type: string
      - description: Only transactions created at or before this RFC3339 timestamp
        in: query
        name: date_to
        type: string
      - description: Max results (default 100)
        in: query
        name: limit
        type: integer
      - description: Skip N results
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/fmis-api_internal_schemas.StockTransaction'
            type: array
        "400":
          description: Invalid filter values
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Missing or invalid token
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List stock transactions
      tags:
      - inventory
...
securityDefinitions:
  BearerAuth:
    description: 'Pass the JWT access token as `Authorization: Bearer <token>`.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"